Splitting main.rs into modules
Restructure main.rs into gui.rs and config.rs for maintainability.
